pdv/django-poc/pyproject.toml
[tool.poetry]
name = "inout"
version = "0.1.0"
description = "Ponto de Vendas"
authors = ["Glayson Olivieri"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1"
djangorestframework = "3.15.2"
argon2-cffi = "23.1.0"
psycopg = {version = "3.2.1", extras = ["binary"]}
django-cors-headers = "^4.4.0"
django-extensions = "^3.2.3"
bpython = "^0.24"
pyjwt = "2.9.0"
pylint = "^3.2.6"
pylint-django = "^2.5.5"
flake8 = "^7.1.1"
djangorestframework-dataclasses = "1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
